+++ test/test_unpark.py
@@ -21,7 +21,9 @@ from freeclimb.model.get_digits import GetDigits
 from freeclimb.model.get_speech import GetSpeech
 from freeclimb.model.hangup import Hangup
 from freeclimb.model.out_dial import OutDial
+from freeclimb.model.park import Park
 from freeclimb.model.pause import Pause
+from freeclimb.model.percl_command import PerclCommand
 from freeclimb.model.play import Play
 from freeclimb.model.play_early_media import PlayEarlyMedia
 from freeclimb.model.record_utterance import RecordUtterance
@@ -35,6 +37,7 @@ from freeclimb.model.set_talk import SetTalk
 from freeclimb.model.sms import Sms
 from freeclimb.model.start_record_call import StartRecordCall
 from freeclimb.model.terminate_conference import TerminateConference
+from freeclimb.model.unpark import Unpark
 globals()['AddToConference'] = AddToConference
 globals()['CreateConference'] = CreateConference
 globals()['Dequeue'] = Dequeue
@@ -43,7 +46,9 @@ globals()['GetDigits'] = GetDigits
 globals()['GetSpeech'] = GetSpeech
 globals()['Hangup'] = Hangup
 globals()['OutDial'] = OutDial
+globals()['Park'] = Park
 globals()['Pause'] = Pause
+globals()['PerclCommand'] = PerclCommand
 globals()['Play'] = Play
 globals()['PlayEarlyMedia'] = PlayEarlyMedia
 globals()['RecordUtterance'] = RecordUtterance
@@ -57,11 +62,12 @@ globals()['SetTalk'] = SetTalk
 globals()['Sms'] = Sms
 globals()['StartRecordCall'] = StartRecordCall
 globals()['TerminateConference'] = TerminateConference
+globals()['Unpark'] = Unpark
 
-from freeclimb.model.percl_command import PerclCommand  # noqa: E501
+from freeclimb.model.unpark import Unpark  # noqa: E501
 
-class TestPerclCommand(unittest.TestCase):
-    """PerclCommand unit test stubs"""
+class TestUnpark(unittest.TestCase):
+    """Unpark unit test stubs"""
 
     def setUp(self):
         pass
@@ -69,10 +75,10 @@ class TestPerclCommand(unittest.TestCase):
     def tearDown(self):
         pass
 
-    def testPerclCommand(self):
-        """Test PerclCommand"""
+    def testUnpark(self):
+        """Test Unpark"""
         # FIXME: construct object with mandatory attributes with example values
-        # model = PerclCommand()  # noqa: E501
+        # model = Unpark()  # noqa: E501
         pass